requirements.txt
pytest
selene>=2.0.b4
webdriver-manager>=3.8.4
allure-pytest